Fixed some issues with the Fill class.
