+++ src/main/scala/org/economicsl/agora/markets/auctions/mutable/continuous/BuyerPostedPriceAuction.scala
@@ -15,6 +15,7 @@ limitations under the License.
 */
 package org.economicsl.agora.markets.auctions.mutable.continuous
 
+import org.economicsl.agora.markets.Fill
 import org.economicsl.agora.markets.auctions.mutable.orderbooks
 import org.economicsl.agora.markets.tradables.{LimitPrice, Price, Quantity}
 import org.economicsl.agora.markets.tradables.orders.Persistent
@@ -28,8 +29,29 @@ import org.economicsl.agora.markets.tradables.orders.bid.BidOrder
   * @tparam B the type of `BidOrder` instances that are stored in the `OrderBook`.
   */
 class BuyerPostedPriceAuction[A <: AskOrder with LimitPrice with Quantity, BB <: orderbooks.BidOrderBook[B], B <: BidOrder with LimitPrice with Persistent with Quantity]
-                             (orderBook: BB, matchingRule: (A, BB) => Option[B], pricingRule: (A, B) => Price)
-  extends PostedPriceAuction[A, BB, B](orderBook, matchingRule, pricingRule)
+                             (protected val orderBook: BB, matchingRule: (A, BB) => Option[B], pricingRule: (A, B) => Price)
+  extends PostedPriceAuctionLike[B, BB] {
+
+  final def fill(order: A): Option[Fill] = {
+    val matchingOrders = matchingRule(order, orderBook) // eventually this will return an iterable?
+    matchingOrders.foreach(matchingOrder => orderBook.remove(matchingOrder.issuer)) // SIDE EFFECT!
+    matchingOrders.map { matchingOrder =>
+      val price = pricingRule(order, matchingOrder)
+      val quantity = math.min(order.quantity, matchingOrder.quantity) // not dealing with residual orders!
+      performance.addValue(surplus(order, matchingOrder))  // SIDE EFFECT!
+      new Fill(order, matchingOrder, price, quantity)
+    }
+  }
+
+  /** Return the surplus generated by a trade.
+    *
+    * @param askOrder an instance of `AskOrder with LimitPrice with Quantity`.
+    * @param bidOrder an instance of `BidOrder with LimitPrice with Quantity`.
+    * @return surplus as measured by the difference between `bidOrder` and `askOrder` `limit` prices.
+    */
+  protected def surplus(askOrder: A, bidOrder: B): Double = bidOrder.limit.value - askOrder.limit.value
+
+}
 
 
 object BuyerPostedPriceAuction {
