+++ src/main/scala/org/economicsl/agora/markets/auctions/mutable/continuous/SellerPostedPriceAuction.scala
@@ -30,6 +30,7 @@ limitations under the License.
 */
 package org.economicsl.agora.markets.auctions.mutable.continuous
 
+import org.economicsl.agora.markets.Fill
 import org.economicsl.agora.markets.auctions.mutable.orderbooks
 import org.economicsl.agora.markets.tradables.{LimitPrice, Price, Quantity}
 import org.economicsl.agora.markets.tradables.orders.Persistent
@@ -43,8 +44,29 @@ import org.economicsl.agora.markets.tradables.orders.bid.BidOrder
   * @tparam A the type of `AskOrder` instances that are stored in the `AskOrderBook`.
   */
 class SellerPostedPriceAuction[B <: BidOrder with LimitPrice with Quantity, AB <: orderbooks.AskOrderBook[A], A <: AskOrder with LimitPrice with Persistent with Quantity]
-                              (orderBook: AB, matchingRule: (B, AB) => Option[A], pricingRule: (B, A) => Price)
-  extends PostedPriceAuction[B, AB, A](orderBook, matchingRule, pricingRule)
+                              (protected val orderBook: AB, matchingRule: (B, AB) => Option[A], pricingRule: (B, A) => Price)
+  extends PostedPriceAuctionLike[A, AB] {
+
+  final def fill(order: B): Option[Fill] = {
+    val matchingOrders = matchingRule(order, orderBook) // eventually this will return an iterable?
+    matchingOrders.foreach(matchingOrder => orderBook.remove(matchingOrder.issuer)) // SIDE EFFECT!
+    matchingOrders.map { matchingOrder =>
+      val price = pricingRule(order, matchingOrder)
+      val quantity = math.min(order.quantity, matchingOrder.quantity) // not dealing with residual orders!
+      performance.addValue(surplus(matchingOrder, order))  // SIDE EFFECT!
+      new Fill(matchingOrder, order, price, quantity)
+    }
+  }
+
+  /** Return the surplus generated by a trade.
+    *
+    * @param askOrder an instance of `AskOrder with LimitPrice with Quantity`.
+    * @param bidOrder an instance of `BidOrder with LimitPrice with Quantity`.
+    * @return surplus as measured by the difference between `bidOrder` and `askOrder` `limit` prices.
+    */
+  protected def surplus(askOrder: A, bidOrder: B): Double = bidOrder.limit.value - askOrder.limit.value
+
+}
 
 
 object SellerPostedPriceAuction {
